Backend/requirements.txt
annotated-types==0.6.0
anyio==4.2.0
boto3==1.34.11
botocore==1.34.11
click==8.1.7
exceptiongroup==1.2.0
fastapi==0.108.0
h11==0.14.0
idna==3.6
jmespath==1.0.1
pydantic==2.5.3
pydantic_core==2.14.6
PyMySQL==1.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
s3transfer==0.10.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.24
starlette==0.32.0.post1
typing_extensions==4.9.0
urllib3==1.26.18
uvicorn==0.25.0
